p2DynArray/List removed, some warnings cleared

+++ Project/Dev_class11_handout/Motor2D/j1HUD.cpp
@@ -65,7 +65,7 @@ bool j1HUD::Update(float dt)
 	App->render->Blit(atlas, -App->render->camera.x + Life_Label_x_pos, -App->render->camera.y + Life_Label_y_pos, &Life_Label, 1.0f, 0, INT_MAX, INT_MAX, false);
 	
 	//Calculate initial position
-	int hearts_pos_x = App->win->GetWindowWHalf() - Heart_Container.w * App->player->hearts_containers_test_purpose * 0.5f - space_between_hearts * (App->player->hearts_containers_test_purpose * 0.5f - 0.5f);
+	int hearts_pos_x = (int)(App->win->GetWindowWHalf() - Heart_Container.w * App->player->hearts_containers_test_purpose * 0.5f - space_between_hearts * (App->player->hearts_containers_test_purpose * 0.5f - 0.5f));
 
 	//Blit heart containers
 	int heart_container_pos_x = hearts_pos_x;
@@ -115,7 +115,7 @@ bool j1HUD::Update(float dt)
 	SDL_Texture* texture_to_blit = App->font->Print(scene_text, { 255,255,255,255 }, App->font->Triforce48);
 	int texture_to_blit_w, texture_to_blit_h = 0;
 	App->font->CalcSize(scene_text, texture_to_blit_w, texture_to_blit_h, App->font->Triforce48);
-	App->render->Blit(texture_to_blit, - App->render->camera.x + PlaceLabelBack.w * 0.5f - texture_to_blit_w * 0.5f, - App->render->camera.y + Window_H - PlaceLabelBack.h * 0.5f, nullptr, 1.0f, 0, INT_MAX, INT_MAX, false, 255);
+	App->render->Blit(texture_to_blit, (int)(- App->render->camera.x + PlaceLabelBack.w * 0.5f - texture_to_blit_w * 0.5f), - App->render->camera.y + Window_H - PlaceLabelBack.h * 0.5f, nullptr, 1.0f, 0, INT_MAX, INT_MAX, false, 255);
 	SDL_DestroyTexture(texture_to_blit);
 
 	return true;
